Reorganise the `Tools/clinic` directory
