+++ Tools/clinic/run_clinic.py
@@ -1,0 +1,4 @@
+from libclinic.cli import main
+
+if __name__ == "__main__":
+    main()
